requirements.txt
requests==2.31.0
python-dotenv==1.0.0
librosa==0.10.1
numpy==1.26.3
tensorflow>=2.16.1
tensorflow-hub>=0.16.1
